fix broken TestTaskCreateThrottling

+++ helix-core/src/test/java/org/apache/helix/task/TestTaskCreateThrottling.java
@@ -42,7 +42,7 @@ public class TestTaskCreateThrottling extends TaskTestBase {
   @Test
   public void testTaskCreatingThrottle() {
     Workflow flow = WorkflowGenerator
-        .generateDefaultRepeatedJobWorkflowBuilder("hugeWorkflow", _driver._configsLimitation + 1)
+        .generateDefaultRepeatedJobWorkflowBuilder("hugeWorkflow", (int) _driver._configsLimitation + 1)
         .build();
     try {
       _driver.start(flow);
